Add details about beforeEach
I was getting errors for `Attempting to inject an unknown injection: 'service:window'` in my unit tests and it was not apparent to me that it was required to do the window mock. I thought it was only if you wanted to test it. Figured adding these details might help others.