Add tests for (nested) setupWindowMock calls
This adds better test coverage for calls to `setupWindowMock()`, specifically the nested usage discussed with @rwjblue in #191#discussion_r454379796.

It turned out that the current implementation (after being fixed by @alexlafroscia in #201) already supports the nested case properly, as the `mockHandler` is set up in the `beforeEach` hook, so as long as there is any nested `setupWindowMock()` the window object will be mocked, and only when a test is not somehow wrapped in a setup call will it see the unmocked window.